pyproject.toml
[tool.poetry]
name = "where-space-backend"
version = "0.1.0"
description = ""
authors = ["Heyuan Zeng <hi@universe.observer>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.0.3"
channels = {extras = ["daphne"], version = "^4.0.0"}
psycopg = {extras = ["binary"], version = "^3.1.18"}
djangorestframework = "^3.14.0"
django-filter = "^24.1"
pillow = "^10.2.0"
django-cors-headers = "^4.3.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
markdown = "^3.5.2"
black = "^24.2.0"
pyyaml = "^6.0.1"
coreapi = "^2.3.3"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = ["ASYNC", "I", "ICN"]
ignore = []

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
